@@ -3,70 +3,24 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
-    "@svgr/webpack": "^5.5.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.11.56",
-    "@types/react": "^17.0.48",
-    "@types/react-dom": "^17.0.17",
-    "babel-jest": "^27.5.1",
-    "babel-plugin-named-asset-import": "^0.3.8",
-    "babel-preset-react-app": "^10.0.1",
-    "bfj": "^7.0.2",
-    "browserslist": "^4.21.3",
-    "camelcase": "^6.3.0",
-    "case-sensitive-paths-webpack-plugin": "^2.4.0",
-    "classnames": "^2.3.1",
-    "css-minimizer-webpack-plugin": "^3.4.1",
-    "dotenv": "^10.0.0",
-    "dotenv-expand": "^5.1.0",
-    "eslint": "^8.23.0",
-    "eslint-config-react-app": "^7.0.1",
-    "eslint-webpack-plugin": "^3.2.0",
-    "file-loader": "^6.2.0",
-    "fs-extra": "^10.1.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-resolve": "^27.5.1",
-    "jest-watch-typeahead": "^1.1.0",
-    "mini-css-extract-plugin": "^2.6.1",
-    "postcss": "^8.4.16",
-    "postcss-flexbugs-fixes": "^5.0.2",
-    "postcss-loader": "^6.2.1",
-    "postcss-normalize": "^10.0.1",
-    "postcss-preset-env": "^7.8.0",
-    "prompts": "^2.4.2",
-    "react": "^17.0.2",
-    "react-app-polyfill": "^3.0.0",
-    "react-dev-utils": "^12.0.1",
-    "react-dom": "^17.0.2",
-    "react-refresh": "^0.11.0",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
-    "resolve": "^1.22.1",
-    "resolve-url-loader": "^4.0.0",
-    "sass": "^1.54.5",
-    "sass-loader": "^12.6.0",
-    "semver": "^7.3.7",
-    "source-map-loader": "^3.0.1",
-    "style-loader": "^3.3.1",
-    "tailwindcss": "^3.1.8",
-    "terser-webpack-plugin": "^5.3.5",
     "typescript": "^4.8.2",
-    "web-vitals": "^2.1.4",
-    "webpack-manifest-plugin": "^4.1.1",
-    "workbox-webpack-plugin": "^6.5.4"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js",
-    "eject": "react-scripts eject",
-    "dev": "NODE_ENV=development BABEL_ENV=development webpack serve --config config/webpackDevServer.config.js",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -85,22 +39,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@babel/core": "^7.18.13",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/preset-env": "^7.18.10",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.18.6",
-    "babel-loader": "^8.2.5",
-    "css-loader": "^6.7.1",
-    "edit-json-file": "^1.7.0",
-    "gh-pages": "^3.2.3",
-    "html-webpack-plugin": "^5.5.0",
-    "prompt": "^1.3.0",
-    "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0",
-    "webpack-dev-server": "^4.10.0",
-    "webpack-log": "^3.0.2"
   }
 }
